The Basics
Outline the basic structure of the FireflyLED

Constructor to initialize state
on() and off() to manage turning the blinking firefly on and off
update(), which is where everything actually happens

Since we don't want to use delays to handle timing, we are using the
millis() method to keep track of time. Right now the FireflyLED is still
pretty simple -- when enabled it simply turns on and off every second.