Refactor into Abstract Base class
Create an abstract base class of FireflyType which wraps up the
functionality around managing the LED pin and whether the firefly is
enabled. Then concrete classes implement the update interface to define
particular firefly algorithms. The first concrete class is the
BlinkingFirefly which wraps the simple on/off blink pattern.

Update the README and add an example.ino class.